Set ubuntu 20.04 as new base image
We would at some point need to set this as the default.
To prepare for that, I think we should make this an option
that is available, and run our tests on this too.

The R stack seems possibly currently unavailable for 20.04
(see https://community.rstudio.com/t/rstudio-server-ubuntu-20-04/64188)
but I'll try it out